Keep the compressed text format importable without the schemas
The migration guidance told data migrations to decode these columns with
`decode_compressed_text` from the schemas package, which no migration
imports: doing so binds a migration to the live schema graph. The format
(marker, ceiling, encoder, decoder) now lives in
`zenml.zen_stores.compressed_text`, a leaf module that imports nothing
from ZenML, and the schema module keeps only the column types while
re-exporting the format names.

Decode failures raise `CompressedTextError`, a `RuntimeError`, instead
of `ValueError`: a corrupt row or an envelope written by a newer server
is a problem with the database, and the server now reports it as a 500
rather than a 400 that blames the client's request. The write guard
keeps raising `ValueError`, because that one is a caller error.
